Refactor code to reuse logic in hls and dash and further complete implementation of dash
Now logic properly handles segment templates, updates, timelines, lists, base segments, and a lot more. Pending to review if some mechanism is required to avoid querying information too early (waiting), and see if we want to keep logic as is for too late requests (or want to skip over to the correct value).
Additionally we haven't had a chance to test against a livestream that eventually ends, so that functionality is not implemented/tested either